pyproject.toml
[project]
name = "try-mlflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "mlflow>=3.1.0",
    "pydantic>=2",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.12.2",
]
